Added 'image url'
In the case of an atricle behind a paywill will cause this image url to be NULL.
So we may need to detect articles that have a paywall and ignore them.
Because even if it does not have the image url, it will only allow us to scrape a part of the summary.
So articles behind a paywall are completely useless.